backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-json-logger>=3.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "supabase>=2.18.1",
    "uvicorn>=0.35.0",
    "google-cloud-storage>=2.10.0",
    "pillow>=10.0.0",
    "geoalchemy2>=0.18.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
